refactor(dsg-utils): moving the removeTSVariableDeclares function
