+++ libs/util/code-gen-utils/src/lib/ast/index.ts
@@ -1,1 +1,2 @@
 export * from "./remove-eslint-comments";
+export * from "./remove-typescript-variable-declares";
