+++ libs/util/code-gen-utils/src/lib/ast/remove-typescript-variable-declares/index.ts
@@ -1,0 +1,1 @@
+export * from "./main";
