+++ packages/data-service-generator/src/server/app-module/create-app-module.ts
@@ -2,6 +2,7 @@ import {
   print,
   readFile,
   removeESLintComments,
+  removeTSVariableDeclares,
 } from "@amplication/code-gen-utils";
 import { builders } from "ast-types";
 import {
@@ -16,7 +17,6 @@ import {
   interpolate,
   importNames,
   addImports,
-  removeTSVariableDeclares,
   removeTSIgnoreComments,
   importDeclaration,
   callExpression,
