+++ packages/data-service-generator/src/server/resource/resolver/create-resolver.ts
@@ -3,6 +3,7 @@ import {
   print,
   readFile,
   removeESLintComments,
+  removeTSVariableDeclares,
 } from "@amplication/code-gen-utils";
 import { builders, namedTypes } from "ast-types";
 import { camelCase } from "camel-case";
@@ -27,7 +28,6 @@ import {
   importNames,
   addAutoGenerationComment,
   addImports,
-  removeTSVariableDeclares,
   removeTSInterfaceDeclares,
   removeTSClassDeclares,
   getClassDeclarationById,
